requirements.txt
pyyaml==5.4.1
ryu==4.34
eventlet==0.30.2
scapy==2.5.0
gnocchiclient==7.0.8
